% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avengertweets.R
\docType{data}
\name{avengertweets}
\alias{avengertweets}
\title{Avengers Endgame Tweets}
\format{
A data frame with 15000 rows and 2 variables:
\describe{
\item{tweet}{\code{character vector} containing user tweets}
\item{id}{ID column}
...
}
}
\source{
Kavita Lolayekar. Twitter Dataset - #AvengersEndgame. \url{https://www.kaggle.com/kavita5/twitter-dataset-avengersendgame/version/1}
}
\usage{
avengertweets
}
\description{
Scraped twitter tweets containing the trending topic #Avengers Endgame
}
\keyword{datasets}
