% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_text_clean.R
\name{twitter_text_clean}
\alias{twitter_text_clean}
\title{Twitter Text Clean: Pre-processing functions for a twitter text corpus}
\usage{
twitter_text_clean(x, dtm = TRUE, ...)
}
\arguments{
\item{x}{\emph{A character vector} containing the twitter text intended to be processed. NA values and empty entries are automatically removed.}

\item{dtm}{A logical argument to specify if the output should be converted to a Document Term Matrix. The {\code{\link[textmineR]{CreateDtm}}} function from the \code{textmineR} library is called here. This is \code{TRUE} by default}

\item{...}{Other arguments to be passed to {\code{\link[textmineR]{CreateDtm}}}}
}
\value{
Processed character vector \emph{or} Document Term Matrix (a format commonly used by many text processing libraries)
}
\description{
A wrapper function for {\code{\link[textmineR]{CreateDtm}}} of the \code{textmineR} library to pre-process twitter data. Further text mining methods can then be conveniently applied.
}
\details{
Twitter data extracted from the web using the Twitter API contains the \code{"at"} symbol due to the formatting of a twitter handle. Additionally, it contains \code{RT} for re-tweets. A \code{https} link to the tweet is also typically present.

\code{twitter_text_clean} conveniently detects and removes twitter handles, links and retweet mentions from text corpuses.
The function outputs a processed character vector. Typical text pre-processing such as removing punctuation, white space, numeric digits and upper case letters is done by default as well.
Alternatively, the function can also output a Document Term matrix. Similar text pre-processing is also additionally done by default here through the excellent {\code{\link[textmineR]{CreateDtm}}} function
}
\examples{
data(avengertweets)

# Outputting a character vector
# Specify dtm = FALSE if the intended output is a character vector
head(twitter_text_clean(avengertweets$tweet, dtm = FALSE))

# Converting into a document-term matrix
# Can pass on additional arguments to CreateDtm
twitter_text_clean(avengertweets$tweet, dtm = TRUE, doc_names = avengertweets$id, cpus = 2)

}
